+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String , Boolean, Enum, Text
+from sqlalchemy import Column, String , Boolean, Enum, Text, Float
 from core.base import Base
 class  User(Base):
     __tablename__ = "Users"
@@ -6,7 +6,7 @@ class  User(Base):
     password = Column(String(500) , nullable = False)
     name = Column(String(50), nullable = False)
     role = Column(Enum('BACKEND', 'FRONTEND'))
-    generation = Column(Integer)
+    generation = Column(Float)
     activity_status = Column(Boolean, default=1, nullable = False)
     admin_status = Column(Boolean, default=0, nullable = False)
     profile_image_url = Column(String(255), nullable=True)
